cmd/compile: initialize variables in the right order
The spec advertises the sort order for variable initialization as
declaration (i.e., lhs of an assignment) and then dependency, but the
compiler currently orders initialization primarily by reference (whether
lhs or rhs of an assignment). To avoid a significant refactor of the
initreorder call chain, this CL introduces data structures to capture
additional information under initreorder, which are then used back up in
initfix to properly order its output.

Unfortunately, this CL results in a performance penalty for make.bash:
it takes half-again as long to run with as without.

Fixes #22326

Change-Id: Iabd07c37b83cc0b080229aa83d7cb5a9780508ee