cmd/compile: initialize variables in the right order
The spec advertises the sort order for variable initialization as
declaration (i.e., lhs of an assignment) and then dependency, but the
compiler currently orders initialization by reference (whether lhs or
rhs of an assignment) and then dependency. This CL introduces NodeSet
and Deps structs in sinit.go, uses instances of these structs in the
call chain under initreorder to track dependency information, and uses
this dependency information back in initfix to properly order its
output.

Fixing the order of the output of initfix ran afoul of a complex bug in
the build toolchain that I did not have time to fully comprehend. This
CL works around it by constructing obj.LinkArches at run time instead of
compile time.

Unfortunately, this CL results in a performance penalty for make.bash:
it takes half-again as long to run with as without.

Aaaaaaand it also hoses the standard library. 😞

Updates #22326

Change-Id: Iabd07c37b83cc0b080229aa83d7cb5a9780508ee