cmd/compile: modify swt.go to avoid double walk of switch
A compiler instrumented to catch any double walks of statements
turned up one example, which could be turned into a double
walk of a switch statement (which generates incorrect code
if not caught).

Fix this by not prepending an already-walked node to the init
list until after the other nodes in the list have been walked.

Fixes #29562.

Change-Id: I0667e8769aba4c3236666cd836a934e256c0bfc5