runtime: fix lock order inversion deadlock in SetMaxHeap
This change fixes a lock inversion deadlock bug in the SetMaxHeap API.

Consider a goroutine (G1) waiting on a notification channel passed to
SetMaxHeap. When it starts to wait on the channel, it'll grab the
channel's lock and may allocate a sudog object, potentially requiring
it to acquire the heap lock.

Now consider another goroutine (G2) which acquires the heap lock and
calls gcSetTriggerRatio. gcSetTriggerRatio may cause a non-blocking
send to occur on the notification channel, which will acquire the
channel lock.

If G2 acquires the heap lock between when G1 acquires the notification
channel's lock but before G1 allocates the sudog, then we have a
deadlock since each goroutine wants to acquire a lock the other has.

This change fixes the problem by decoupling actually sending messages
through the notification channel from determining whether the GC should
notify the program at all. Previously, this logic was captured in
closure which would get executed with the heap lock held. In this
change, gcSetTriggerRatio now returns a Boolean value indicating
whether a change in GC policy was made. If true, the runtime sends a
message on the notification channel once the heap lock is released.

Generally, the heap lock is released shortly after gcSetTriggerRatio is
called (except during STW, at which point we don't need to worry about
locks) so there shouldn't be a significant delay between making the
decision to send a notification and sending the notification itself.

Change-Id: I6efbc97d6ed52456a79a3be494581168243192e6