runtime: fix lock order inversion deadlock in SetMaxHeap
This change fixes a lock inversion deadlock bug in the SetMaxHeap API.

Consider a goroutine (G1) waiting on a notification channel passed to
SetMaxHeap. It may grab the channel's lock, which may allocate a sudog
object, potentially requiring it to acquire the heap lock.

Now consider another goroutine (G2) which acquires the heap lock and
calls gcSetTriggerRatio. gcSetTriggerRatio may cause a non-blocking
send to occur on the notification channel.

If G2 acquires the heap lock between when G1 acquires the notification
channel's lock but before G1 allocates the sudog, then we have a
deadlock since each goroutine wants to acquire a lock the other has.

Change-Id: I6efbc97d6ed52456a79a3be494581168243192e6