Update patch set 1
Patch Set 1:

> > I think we also need to do something about the panic, but that's a separate issue. LGTM.
> 
> I'm sorry; I did a poor job of reading this change.
> 
> This adds the test name to the panic, but it doesn't address the problem that log messages between a test completing and the last test completing are silently lost. The window for the panic is so small as to make it highly racy and (IMO) not useful as it stands.

Right, this is meant to be a minimal change for 1.12: let's at least report the right message for the cases where it does panic.

I agree we shouldn't lose log messages, but that is not a new problem, and I think we shouldn't try to tackle it at this point in the release cycle.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>