Update patch set 9
Patch Set 9:

Brian, Baokun, sorry for the bump, but is it clear which person this is pending at this point?

Is this still pending an integration test, or is the current level of testing considered sufficient at this point?

If there is still a desire for a shorter test example, I think this is a shorter example:
https://gist.github.com/artyom/3f73ca859f7476902ba4e1c78da0429d

(from the related #24087)

Patch-set: 9
CC: Gerrit User 28792 <28792@62eb7196-b449-3ce5-99f1-c037f21e1705>