Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> If vscode is broken then please raise an issue at that project, thanks.

I think the original intention here was to use the ``<>'' convention which godoc recognizes converts to “<>”.

We can rectify that or directly use the unicode quotes. But changing it to double backticks doesn't help because godoc doesn't recognize it either and the docs will still remain broken.

Patch-set: 1
CC: Gerrit User 24096 <24096@62eb7196-b449-3ce5-99f1-c037f21e1705>