Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Please just remove the quotes entirely.  We don't use quotes in godoc comments.  Thanks.

OTOH, "up to a total length of bits bits" is pretty awkward, which is probably why we violated our normal no-quotes rule in this case. (It's also a really old comment.)

I'd probably just reword it so it's less awkward and doesn't need quotes, even if that involves changing the parameter names.

Patch-set: 1