Update patch set 1
Patch Set 1:

(2 comments)

Do we understand the performance implications of this? Maybe it's not too bad because this is only triggered if we've already gone to the trouble of scavenging the span.

It worries me a little bit that we don't track over-scavenging. For example, if the code in this CL causes various little sysUnuseds, but every call to scavengeLargest picks the largest available span to scavenge, we could be scavenging far more than necessary. If scavenging in excess of the nbytes passed to scavengeLargest were tracked as credit, then later calls to scavengeLargest could draw down that credit without doing any extra scavenging.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>