cmd/compile: avoid confusing error on pkg.keyword
The program below:

	import "go/types"
	func _() {
		var _ types.const
	}

Used to error twice, where the first error seems to be malformed:

	./f.go:6:8: cannot refer to unexported name types._
	./f.go:6:14: syntax error: unexpected const, expecting name

Instead, remove that error and only show the syntax error, which is
what's useful.

First, we need to teach the syntax parser to emit empty names in the
case of syntax errors, and not "_" names, to differentiate
"types.<keyword>" from a syntactically correct "types._".

Then, add a check in subr.go so that we don't emit the extra error if
the name was empty, since we already reported a syntax error.

While at it, add a few more test cases involving package selectors with
underscore names, to ensure we don't break those errors.

Fixes #30039.

Change-Id: I03e27a17ed38dc28250f68790351e63db594e8fe