runtime: prefer unscavenged spans over scavenged spans
Currently when allocating, we try to reuse free spans. Our policy for
picking a span to use is to choose the best fit across both scavenged
and unscavenged spans. There are, however, a number of reasons not to do
this.

Firstly, scavenged spans incur a cost upon access because even if these
spans are actually backed by physical pages, they will still incur a
soft page fault on access. Secondly, allocating out of a scavenged span
forces us to go and scavenge other memory to make up for the RSS growth,
which also incurs a cost.

By preferring unscavenged spans, we can minimize those costs, and rely
on the scavenging mechanisms to pick spans which are less likely to be
reused.

For #30333.

Change-Id: I7f3c0b47af743674a5192532e1d758f8a9a13099