Update patch set 1
Patch Set 1:

Note that some people want to be able to look at a .a file and figure out which packages it imports.  As far as I know, they can currently do that by running `go tool nm` and looking for .init symbols.  At first glance that will no longer work after this CL.  I think that for better or for worse we need to document a way for people to do this that will work with both old and new versions of Go.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>