Update patch set 1
Patch Set 1:

(1 comment)

Thanks for your change and apologies for the delay.

The primary question I have is whether JS modules are out of draft spec mode or not.

Python has added support for the .mjs extension as a reference (https://bugs.python.org/issue31715). But this discussion seems better on an issue.

Thanks

Patch-set: 1
CC: Gerrit User 22285 <22285@62eb7196-b449-3ce5-99f1-c037f21e1705>