cmd/link: add infrastructure for interning symbol strings
DO NOT REVIEW
DO NOT SUBMIT

Introduce some new infrastructure designed to help with interning
linker symbol strings. This includes a new type sym.SymName (an opaque
type encapsulating the name given to a sym.Symbol) and SymNameTable
(table that maintains string storage for a set of symbol names).

The basic idea here is to split each symbol name N into two pieces
{ X Y }, where X contains the prefix of N up to and including the first
"." character (or is empty if N does not contain a ".") and Y contains
the remaining characters in N following the ".". Each of these
pieces/fragments is inserted into the string table, yielding an
index/token: we hash the string then keep a map from hash code to a
slot containing a vector of strings.

Change-Id: Ib6709693cd9bbe0c8067543ffff612e5ea3b7bb7