Update patch set 2
Patch Set 2:

In isolation, I understand this change, but it makes Reader.Reset and Writer.Reset asymmetrical, since Writer.Reset still doesn't permit taking in nil. Also, using nil to signal reusing the underlying reader is a somewhat odd API.

As you mentioned in the issue, there is a workaround for this. The documentation for Reader.Multistream currently states:

> If the underlying reader implements io.ByteReader, it will be left positioned just after the gzip stream. To start the next stream, call z.Reset(r) followed by z.Multistream(false). If there is no next stream, z.Reset(r) will return io.EOF.

Thus, it explicitly calls out the fact that users of this feature need to provide an io.ByteReader to properly use this feature.

Patch-set: 2
CC: Gerrit User 9735 <9735@62eb7196-b449-3ce5-99f1-c037f21e1705>