Update patch set 2
Patch Set 2:

I thought about this some more and I'm more hesitant to use Reset(nil) to signal that the underlying reader should be re-used. Seeing that used in code does not make it clear what's happening without reading the documentation. In fact, it's the opposite of what the call seems to be doing.

Furthermore, this would make gzip different from the other compress packages where Reset(nil) explicitly clears the underlying reader and is sometimes done to help the Go GC be able to reclaim the memory for the underlying reader. This is useful if a compression Reader is cached within a larger data structure.

> it also doesn't explicitly mention that the underlying io.Reader is wrapped and this behavior is expected. It does not state that passing the original io.Reader to Reset will not work UNLESS it implements ByteReader.

Actually, it does. The documentation for gzip.NewReader says:

> If r does not also implement io.ByteReader, the decompressor may read more data than necessary from r.

This implies some form of wrapping or internal buffering.

> Open to other options...

Alternatives:
* A gzip.ReuseReader sentinel reader value that can be passed to Reset to explicitly signal that the underlying reader should be reused. Thus, this would be more readable as it would read as Reset(gzip.ReuseReader). Even with this, iterating over each stream is still clunky, since you have to call Reader.Multistream(false) after each Reset.
* An additional Reader.NextStream method that is explicitly called to advance to the next stream. This is better than the previous option since it doesn't require calling Reader.Multistream(false) after every stream.

Overall, I'm inclined to make no changes. In my opinion, dealing with individual GZIP files in a multistream file is fundamentally clunky and there is a workaround as you noted in the issue.

Patch-set: 2