Update patch set 2
Patch Set 2:

> If clunky is ok, and you don't want changes then I'll just fork it instead of wasting your time further. Cheers

I never said this was a waste of time.

API design is a process when multiple competing goals are being balanced. You have a legitimate use-case and the current API is clunky for that use-case. It's also a use-case that is arguably a small proportion of what people use the gzip package for. We could add more API to make this slightly less sharp. However, the cost of adding API here is more cognitive load for the common case where users don't care about this feature.

I considered your proposal and several alternatives, and I believe keeping the API as is a better state when considering the entire ecosystem. We want to be careful not to overfit an API towards some narrow set of use-cases. If we were to re-design gzip from scratch, there a few things I would love to change, the API for multi-streaming may be one of them.

Patch-set: 2