Update patch set 1
Patch Set 1:

(1 comment)

Thanks for this improvement.

I think the actual value used to grow the StringBuilder needs some
justification. With your value of 128 the test cooke can be written
in one alloc but a lot of cookies have much larger cookie values
and will trigger additional allocations: E.g. a 1k long value will
trigger 3 allocs with your b.Grow(128) but only 4 allocs without
pre-growing the StringBuilder. For such long cookie values values
(these are common!) this CL has not much effect.

On the other hand 128 is pretty large for "small" cookies, especially
those with just short Name and Value. Here 128 seems a bit wasteful.

Thus said I think that yes, the StringBuilder should be pre-grown
but the actual value should be a reasonable value which optimizes
the "average" use, not just the single cookie used in the benchmark.

Maybe the initial amount could even be based on the length of Name
and Value and the presence of other fields?

Patch-set: 1
CC: Gerrit User 5050 <5050@62eb7196-b449-3ce5-99f1-c037f21e1705>