Update patch set 1
Abandoned

Thanks for the contribution, but no one should be using a cipher.Block directly (it will only encrypt the first 16 bytes! silently! with no nonce!), so we will not add an example for it. If you think there should be an example in this package, show how to use it with cipher.NewGCM.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon