Update patch set 12
Patch Set 12:

This breaks ppc64 when compiled with PIC. I think the problem is that for the new case you really only need to insert a single NOP, but ginsnop2 was used which isn't really right for this situation since it will load r2 from a location that might not have been initialized. The previous ginsnop2 was a special case for deferreturn, not a general function for inserting a NOP in all situations.

Patch-set: 12
CC: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>